capstone/requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
numpy>=1.26
tqdm>=4.66
pymupdf>=1.24.0
python-dotenv>=1.0.1
requests>=2.31
faster-whisper>=1.0.3
colorama>=0.4.6
